More updates on send cheeps
