+++ src/Chirp.Core/DomainModel/Author.cs
@@ -1,9 +1,15 @@
-namespace  Chirp.Core.DomainModel ;
+
+namespace Chirp.Core.DomainModel ;
 
 public class Author {
     public int AuthorId {get; set;}
     public String Name { get; set;}
     public String Email { get; set;}
     public ICollection<Cheep> Cheeps { get; set;}
+
+    public static implicit operator Author?(Task<Author>? v)
+    {
+        throw new NotImplementedException();
+    }
 }
