+++ src/Chirp.Razor/Pages/Shared/_Layout.cshtml
@@ -10,9 +10,11 @@
 <div class=page>
     <h1><img src="/images/icon1.png" alt="Icon1"/>Chirp!</h1>
     <partial name="_LoginPartial.cshtml" />
-
-    <partial name="_LoginPartial.cshtml" />
-
+    @if (User.Identity.IsAuthenticated){
+        
+        <partial name="SendCheeps.cshtml"/>
+    }
+    
     <div class=body>
             @RenderBody()
     </div>
